Patient Profile Module server

(For use outside of the DaVinci framework)
Runs the server that populates the UI returned by mod_patient_profile_UI().
Requires a matching call to that function.

Usage

mod_patient_profile_server(
  id,
  subject_level_dataset,
  extra_datasets,
  subjid_var,
  sender_ids,
  summary,
  listings,
  plots
)

Arguments

id

[character] Unique shiny ID. Must match the ID provided to mod_patient_profile_UI().

subject_level_dataset

[reactive(data.frame)] Visit-independent subject information.

extra_datasets

[reactive(data.frame(n))] Visit-dependent subject datasets.

subjid_var

[character] Unique subject identifier column.

sender_ids

[character(n)] (optional) Identifiers of modules allowed to request the display of target patient IDs.

summary

[list] (optional) Subject-level summary section. Composed of:

vars [character(n)] Values to display in the summary section. Indexes into dataset subject_level_dataset_name.

column_count [integer] Column count for the summary section.

listings

[list(n)] (optional) Listings section. Composed of:

dataset [character] Dataset name.

default_vars [character(n)] (optional) Default columns to display. If not specified, the first six columns are selected by default. Indexes into dataset dataset.

plots

[list] (optional) Plot section. Composed of:

x_axis_unit [character] (optional) Defines the time units of the x-axis in the plots. Expects ["weeks"/"days"] values. Defaults to "days".

x_axis_breaks [integer|numeric(n)] (optional) Defines how many breaks will be used in the x-axis. When a single integer is passed it will use base::pretty to compute that number of breaks. If more than one value is passed it will use those breaks in the x-axis (e.g. if c(1, 2, 3) is passed it will show breaks at days/weeks 1, 2 and 3). Defaults to 5.

timeline_info [list] Start and end study dates. Composed of:

icf_date [character] (optional) Informed Consent Form signing Date. Indexes into dataset subject_level_dataset_name. Expects [Date|POSIXt] values.

trt_start_date [character] Treatment Start Date, used also as Day 1 Reference Date. Indexes into dataset subject_level_dataset_name. Expects [Date|POSIXt] values.

trt_end_date [character] Treatment End Date. Indexes into dataset subject_level_dataset_name. Expects [Date|POSIXt] values.

part_end_date [character] (optional) Participation End Date. Indexes into dataset subject_level_dataset_name. Expects [Date|POSIXt] values.

range_plots [list(n)] Plots for range-like events (adverse events, concomitant medications, ...). Composed of:

dataset [character] Dataset name.

vars [list] Variables of interest for the event. Composed of:

start_date [character] Start date. Indexes into dataset dataset. Expects [Date|POSIXt] values.

end_date [character] End date. Indexes into dataset dataset. Expects [Date|POSIXt] values.

decode [character] Descriptive event label. Indexes into dataset dataset. Expects [character|factor] values.

grading [character] (optional) Degree of the event, used for color-coding the plot. Indexes into dataset dataset. Expects [character|factor] values.

serious_ae [character] (optional) Marker of event seriousness, usually reserved for adverse event datasets. Indexes into dataset dataset. Expects [logical|"Y"/"N"] values.

tooltip [character(n)] Block of text to display as hover information over the left edge of each range. The names of this list are included as literal text and honor three basic HTML formatting elements: <b>, <i>, <br>). The columns the values refer to are populated with the value on the dataset relevant to any given row. Indexes into dataset dataset.

value_plots [list(n)] Plots for value-like (lab measurements, vital signs, ...) traces. Composed of:

dataset [character] Dataset name.

vars [list] Variables of interest for the trace. Composed of:

analysis_param [character] Parameter. Indexes into dataset dataset. Expects [character|factor] values.

analysis_val [character] Value. Indexes into dataset dataset. Expects [numeric] values.

analysis_date [character] Date. Indexes into dataset dataset. Expects [Date|POSIXt] values.

analysis_indicator [character] (optional) Analysis reference range indicator. Indexes into dataset dataset. Expects [character|factor] values.

range_low_limit [character] (optional) Lower limit of the reference range. Indexes into dataset dataset. Expects [numeric] values.

range_high_limit [character] (optional) Upper limit of the reference range. Indexes into dataset dataset. Expects [numeric] values.

summary_stats [character] (optional) Additional value column for summary statistics. Indexes into dataset dataset. Expects [numeric] values.

tooltip [character(n)] Block of text to display as hover information over each point of the trace. The names of this list are included as literal text and honor three basic HTML formatting elements: <b>, <i>, <br>). The columns the values refer to are populated with the value on the dataset relevant to any given row. Indexes into dataset dataset.

default_analysis_params [character(n)] (optional) A vector of character values specifying the default analysis parameters (values from the variable specified by analysis_param) to display.

vline_vars [character(n)] Place vertical dashed lines on days indicated by these dataset columns. Indexes into dataset subject_level_dataset_name. Expects [integer|Date|POSIXt] values.

vline_day_numbers [integer(n)] (optional) Place vertical dashed lines on days indicated by this parameter. Represents a CDISC (non-zero) Study Day.

palette [character(n)] (optional) If a name on this list matches the text on a plot element, the associated color will be applied to that element. This mapping takes precedence over the built-in palette. Contains either an HTML (#xxxxxx) or an R color.

See also